# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Ticketrs Please API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://tickets-please.test'
paths:
...
  '/api/v1/tickets/{id}':
    get:
      summary: show
      operationId: show
      description: 'Display the specified resource.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    type: ticket
                    id: 1
                    attributes:
                      title: 'quae sint reprehenderit'
                      description: 'Asperiores eos non minima et libero. Quis magni veniam enim. Rem quam quidem doloribus enim molestiae quisquam nobis.'
                      status: A
                      createdAt: '2024-09-08T07:42:15.000000Z'
                      updatedAt: '2024-09-08T07:42:15.000000Z'
                    relationships:
                      author:
                        data:
                          type: user
                          id: 8
                        links:
                          self: 'http://tickets-please.test/api/v1/authors/8'
                    links:
                      self: 'http://tickets-please.test/api/v1/tickets/1'
                properties:
                  data:
                    type: object
                    properties:
                      type:
                        type: string
                        example: ticket
                      id:
                        type: integer
                        example: 1
                      attributes:
                        type: object
                        properties:
                          title:
                            type: string
                            example: 'quae sint reprehenderit'
                          description:
                            type: string
                            example: 'Asperiores eos non minima et libero. Quis magni veniam enim. Rem quam quidem doloribus enim molestiae quisquam nobis.'
                          status:
                            type: string
                            example: A
                          createdAt:
                            type: string
                            example: '2024-09-08T07:42:15.000000Z'
                          updatedAt:
                            type: string
                            example: '2024-09-08T07:42:15.000000Z'
                      relationships:
                        type: object
                        properties:
                          author:
                            type: object
                            properties:
                              data:
                                type: object
                                properties:
                                  type:
                                    type: string
                                    example: user
                                  id:
                                    type: integer
                                    example: 8
                              links:
                                type: object
                                properties:
                                  self:
                                    type: string
                                    example: 'http://tickets-please.test/api/v1/authors/8'
                      links:
                        type: object
                        properties:
                          self:
                            type: string
                            example: 'http://tickets-please.test/api/v1/tickets/1'
      tags:
        - 'Ticket Management'
    delete:
      summary: destroy
      operationId: destroy
      description: 'Remove the specified resource from storage.'
      parameters: []
      responses: {  }
      tags:
        - 'Ticket Management'
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the ticket.'
        example: 1
        required: true
        schema:
          type: integer
  '/api/v1/tickets/{ticket_id}':
    put:
      summary: replace
      operationId: replace
      description: 'Replace the specified resource in storage.'
      parameters: []
      responses: {  }
      tags:
        - 'Ticket Management'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: object
                  description: ''
                  example: []
                  properties:
                    attributes:
                      type: object
                      description: ''
                      example: []
                      properties:
                        title:
                          type: string
                          description: ''
                          example: magni
                        description:
                          type: string
                          description: ''
                          example: 'Blanditiis amet ad dolor rem perspiciatis aut ullam odio.'
                        status:
                          type: string
                          description: ''
                          example: C
                          enum:
                            - A
                            - C
                            - H
                            - X
                    relationship:
                      type: object
                      description: ''
                      example: []
                      properties:
                        author:
                          type: object
                          description: ''
                          example: []
                          properties:
                            data:
                              type: object
                              description: ''
                              example: []
                              properties:
                                id:
                                  type: integer
                                  description: ''
                                  example: 7
    patch:
      summary: update
      operationId: update
      description: 'Update the specified resource in storage.'
      parameters: []
      responses: {  }
      tags:
        - 'Ticket Management'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: object
                  description: ''
                  example: []
                  properties:
                    attributes:
                      type: object
                      description: ''
                      example: []
                      properties:
                        title:
                          type: string
                          description: ''
                          example: ea
                        description:
                          type: string
                          description: ''
                          example: 'Et architecto officia esse dolorum similique.'
                        status:
                          type: string
                          description: ''
                          example: A
                          enum:
                            - A
                            - C
                            - H
                            - X
                    relationship:
                      type: object
                      description: ''
                      example: []
                      properties:
                        author:
                          type: object
                          description: ''
                          example: []
                          properties:
                            data:
                              type: object
                              description: ''
                              example: []
                              properties:
                                id:
                                  type: string
                                  description: ''
                                  example: null
    parameters:
      -
        in: path
        name: ticket_id
        description: 'The ID of the ticket.'
        example: 1
        required: true
        schema:
          type: integer
...
  '/api/v1/users/{id}':
    get:
      summary: show
      operationId: show
      description: 'Display the specified user.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    type: user
                    id: 1
                    attributes:
                      name: 'Ms. Chloe Marvin'
                      email: lockman.aaron@example.com
                      isManager: true
                    links:
                      self: 'http://tickets-please.test/api/v1/authors/1'
                properties:
                  data:
                    type: object
                    properties:
                      type:
                        type: string
                        example: user
                      id:
                        type: integer
                        example: 1
                      attributes:
                        type: object
                        properties:
                          name:
                            type: string
                            example: 'Ms. Chloe Marvin'
                          email:
                            type: string
                            example: lockman.aaron@example.com
                          isManager:
                            type: boolean
                            example: true
                      links:
                        type: object
                        properties:
                          self:
                            type: string
                            example: 'http://tickets-please.test/api/v1/authors/1'
      tags:
        - 'User Management'
    delete:
      summary: destroy
      operationId: destroy
      description: 'Remove the specified user from storage.'
      parameters: []
      responses: {  }
      tags:
        - 'User Management'
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the user.'
        example: 1
        required: true
        schema:
          type: integer
  '/api/v1/users/{user}':
    put:
      summary: replace
      operationId: replace
      description: 'Update the specified resource in storage.'
      parameters: []
      responses: {  }
      tags:
        - 'User Management'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: object
                  description: ''
                  example: []
                  properties:
                    attributes:
                      type: object
                      description: ''
                      example: []
                      properties:
                        name:
                          type: string
                          description: ''
                          example: sed
                        email:
                          type: string
                          description: 'Must be a valid email address.'
                          example: lexie.skiles@example.net
                        isManager:
                          type: boolean
                          description: ''
                          example: false
                        password:
                          type: string
                          description: ''
                          example: '2-lKyX:YXv&=La1'
    patch:
      summary: update
      operationId: update
      description: 'Update the specified user in storage.'
      parameters: []
      responses: {  }
      tags:
        - 'User Management'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: object
                  description: ''
                  example: []
                  properties:
                    attributes:
                      type: object
                      description: ''
                      example: []
                      properties:
                        name:
                          type: string
                          description: ''
                          example: incidunt
                        email:
                          type: string
                          description: 'Must be a valid email address.'
                          example: fadel.tillman@example.com
                        isManager:
                          type: boolean
                          description: ''
                          example: false
                        password:
                          type: string
                          description: ''
                          example: 'zO^"a}'
    parameters:
      -
        in: path
        name: user
        description: 'The user.'
        example: 1
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []